requirements.txt
Flask
Flask-Cors
Flask-Limiter
flask-sqlalchemy

langchain
langchain-core
langchain-community
langchain-google-genai
langchain-openai
langchain-groq

google-generativeai
openai
huggingface_hub

python-dotenv
Pillow
psycopg2-binary




